rename console timer module
