+++ modules/console-timer.js
